+++ src/main/webapp/WEB-INF/view/artist/modifyArtist.jsp
@@ -166,11 +166,7 @@ prefix="c" %> <%@ page session="false" %>
                   <li class="nav-item">
                     <a class="nav-link" href="/mypage" id="detail">My page</a>
                   </li>
-                  <li class="nav-item">
-                    <a class="nav-link" id="detail" href="/history"
-                      >결제 내역</a
-                    >
-                  </li>
+
                 </c:if>
                 <c:if test="${user.role == '관리자'}">
                   <li class="nav-item">
