+++ src/main/webapp/WEB-INF/view/exhibit/exhibitModify.jsp
@@ -164,11 +164,7 @@ uri="http://www.springframework.org/tags/form"%> <%@ page session="false" %>
                   <li class="nav-item">
                     <a class="nav-link" href="/mypage" id="detail">My page</a>
                   </li>
-                  <li class="nav-item">
-                    <a class="nav-link" id="detail" href="/history"
-                      >결제 내역</a
-                    >
-                  </li>
+
                 </c:if>
                 <c:if test="${user.role == '관리자'}">
                   <li class="nav-item">
